Add one-two sample
This sample was extracted from https://youtu.be/2ZrWHtvSog4 (Youtube:
"1-Minute Audio Test for Stereo Speakers & Headphones") and normalized.

Its purpose is to demonstrate pops that occur in the headphone output.

Related to the discussion here:
https://gist.github.com/hamidzr/dd81e429dc86f4327ded7a2030e7d7d9#gistcomment-3519798

This sample has nothing to do with the patches in this repository, I'm just
piggybacking it on here to host the file.